requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
google-genai==1.39.0
python-dotenv==1.0.0